rename bridger to bridge
